task: move to Vite for builds
This enables us to use Sass and TS for our source files
