+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Changed to [Vite](https://vitejs.dev/) for builds
+
 ## [0.0.3] - 2023-01-11
 
 ### Added
