+++ manifest.json
@@ -1,13 +1,13 @@
 {
   "name": "Am I Talking A Lot?",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "Measure talking times in Google Meet",
   "manifest_version": 3,
   "content_scripts": [
     {
       "matches": ["https://meet.google.com/*"],
-      "js": ["amITalkingALot.js"],
-      "css": ["amITalkingALot.css"]
+      "js": ["dist/aital.js"],
+      "css": ["dist/aital.css"]
     }
   ],
   "icons": {
