+++ src/styles.scss
@@ -1,0 +1,170 @@
+details[data-id="aital"] {
+  background: white;
+  border: 1px solid hsla(0, 0%, 0%, 0.2);
+  border-radius: 8px;
+  left: 16px;
+  position: fixed;
+  top: 16px;
+  width: 240px;
+  z-index: 1;
+
+  summary {
+    display: flex;
+    gap: 32px;
+    justify-content: space-between;
+    padding: 8px;
+
+    svg {
+      height: 16px;
+      width: 16px;
+    }
+  }
+
+  &[open] summary svg:first-of-type {
+    display: none;
+  }
+
+  &:not([open]) summary svg:last-of-type {
+    display: none;
+  }
+
+  > div {
+    align-items: center;
+    display: grid;
+    font-size: smaller;
+    gap: 8px;
+    grid-template-columns: max-content 1fr max-content;
+    padding: 0 8px;
+
+    :nth-last-child(-n + 3) {
+      margin-bottom: 8px;
+    }
+  }
+
+  progress {
+    appearance: none;
+    border-radius: 8px;
+    overflow: hidden;
+    width: 100%;
+
+    &::-webkit-progress-bar {
+      background-color: hsl(0, 0%, 90%);
+    }
+
+    &[data-level="0"]::-webkit-progress-value {
+      background-color: #00ac48;
+    }
+
+    &[data-level="1"]::-webkit-progress-value {
+      background-color: #ffbb00;
+    }
+
+    &[data-level="2"] {
+      animation: aital-little-shake 0.4s linear infinite;
+
+      &::-webkit-progress-value {
+        background-color: #e94337;
+      }
+    }
+
+    &[data-level="3"] {
+      animation: aital-big-shake 0.3s linear infinite;
+
+      &::-webkit-progress-value {
+        background-color: #971a11;
+      }
+    }
+  }
+}
+
+@keyframes aital-little-shake {
+  0% {
+    transform: translate(0);
+  }
+
+  10% {
+    transform: translate(-1px, -1px);
+  }
+
+  20% {
+    transform: translate(1px, -1px);
+  }
+
+  30% {
+    transform: translate(-1px, 1px);
+  }
+
+  40% {
+    transform: translate(1px, 1px);
+  }
+
+  50% {
+    transform: translate(-1px, -1px);
+  }
+
+  60% {
+    transform: translate(1px, -1px);
+  }
+
+  70% {
+    transform: translate(-1px, 1px);
+  }
+
+  80% {
+    transform: translate(-1px, -1px);
+  }
+
+  90% {
+    transform: translate(1px, -1px);
+  }
+
+  100% {
+    transform: translate(0);
+  }
+}
+
+@keyframes aital-big-shake {
+  0% {
+    transform: translate(0);
+  }
+
+  10% {
+    transform: translate(-2px, -2px);
+  }
+
+  20% {
+    transform: translate(2px, -2px);
+  }
+
+  30% {
+    transform: translate(-2px, 2px);
+  }
+
+  40% {
+    transform: translate(2px, 2px);
+  }
+
+  50% {
+    transform: translate(-2px, -2px);
+  }
+
+  60% {
+    transform: translate(2px, -2px);
+  }
+
+  70% {
+    transform: translate(-2px, 2px);
+  }
+
+  80% {
+    transform: translate(-2px, -2px);
+  }
+
+  90% {
+    transform: translate(2px, -2px);
+  }
+
+  100% {
+    transform: translate(0);
+  }
+}
